checkpoint in evo search refactor
